+++ lib/redmine_asap_theme/easy_gantt_helper_patch.rb
@@ -1,19 +1,27 @@
 module RedmineAsapTheme
   module EasyGanttHelperPatch
-    MIN_GANTT_DATE = Date.new(1900, 1, 1)
+    if Redmine::Plugin.installed?(:easy_gantt)
+      MIN_GANTT_DATE = Date.new(1900, 1, 1)
 
-    def api_render_issues(api, issues, **kwargs)
-      issues.each do |issue|
-        issue.start_date = nil if issue.start_date && issue.start_date < MIN_GANTT_DATE
-        issue.due_date   = nil if issue.due_date   && issue.due_date   < MIN_GANTT_DATE
+      def gantt_format_column(entity, column, value)
+        if column.is_a?(QueryRelationsColumn)
+          column.value_object(entity)
+        else
+          super
+        end
       end
-      super
+
+      def api_render_issues(api, issues, **kwargs)
+        issues.each do |issue|
+          issue.start_date = nil if issue.start_date && issue.start_date < MIN_GANTT_DATE
+          issue.due_date   = nil if issue.due_date   && issue.due_date   < MIN_GANTT_DATE
+        end
+        super
+      end
     end
   end
 end
 
-Rails.configuration.to_prepare do
-  if Redmine::Plugin.installed?(:easy_gantt)
-    EasyGanttHelper.prepend RedmineAsapTheme::EasyGanttHelperPatch
-  end
+if Redmine::Plugin.installed?(:easy_gantt)
+  EasyGanttHelper.prepend RedmineAsapTheme::EasyGanttHelperPatch
 end